third_party/blink/web_tests/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/at-container-013-expected.txt

This is a testharness.js-based test.
FAIL Container ineligible for containment assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL Changing containment eligibility invalidates style assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
